Remove Deface hook for Pages in favor of MenuItem
This commit replaces the Deface hook with a standard MenuItem approach to
add the "Pages" entry into the Admin menu.

Besides removing Deface, this change has an additional benefit that the
Pages entry now behaves in the same way that the original Solidus admin
menu entries do.
(Previously, with Deface, the "Pages" entry was being added to the menu
even on the Admin Login page, where the standard menu should remain empty.)