Cleanup gemspec and Gemfile
Any development dependency is now correctly declared in the gemspec
and only addressed in the Gemfile if special options are desired (e.g.
needs to load a specific branch or a different version depending on
the context).

Since the extension requires Solidus 2.3+ any support for older
versions has been removed.